refactor: refactor sim_server & dataset
